feat(file): `FileReader::Reader` is thread-safe, and `FileReader::read_as_bytes` consumes `self`.
First off, `FileReader::Reader` is required to be `Send`, so that the result of
`read_as_bytes` can be sent over thread safely.

Second, `FileReader::read_as_bytes` takes ownership of `self`, so it consumes
the value itself. That way, it's no longer require for the file reader to be
mutable, or to be a reference with a specific lifetime etc.