chore(cargo): Update `unicode-linebreak` manually.
`unicode-linebreak` 0.1.5 has been released. It removes the `build.rs`
script (it's moved in another separate internal crate). Moving to 0.1.5
removes the following dependencies from the tree:

* `ahash`,
* `aho-corasick`,
* `hashbrown`,
* `once_cell`,
* `regex`,
* `regex-syntax`,
* `version_check`,
* `wasi`.

This isn't negligible.

Also, `ahash` was at version 0.7.6 (latest version is —at the time of
writing— 0.8.8), which doesn't compile on Rust nightly. `ahash` was
required by `hashbrown`, which was required as a build-dependency of
`unicode-linebreak` 0.1.4. It's no longer the case in 0.1.5.