feat(object): Simplify section names and symbol names resolution.
This patch removes a large unsafe code surface (the `slice` module
entirely) to replace it with a new `File::fetch_section_names`
method. To make it work, the `Section::name` field now takes an
`Option<BString>` instead of an `Option<Cow<'a, BStr>>` because of the
borrow checker.

This patch also introduces the `File::strings_section` method, and
propagates that to the `Data::symbols` method to automatically fetch the
symbol names.